requirements.txt
# 基础依赖项
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
insightface>=0.6.0
onnxruntime>=1.8.0
scipy>=1.7.0
matplotlib>=3.3.0

# PyQt5依赖项
PyQt5>=5.15.0
PyQt5-tools>=5.15.0
pyqtdarktheme>=2.1.0

# 视频处理依赖项
ffmpeg-python>=0.2.0

# 日志和工具
tqdm>=4.45.0
colorama>=0.4.4 